loktar_init/app/requirements.txt
pyaml==15.8.2